@@ -5,34 +5,48 @@
     <modelVersion>4.0.0</modelVersion>
 
     <name>Squiggly Filter Jackson</name>
-    <url>https://github.com/bohnman/squiggly-filter-jackson</url>
+    <url>https://github.com/jacquant/squiggly-filter-jackson</url>
     <description>The Squiggly Filter is a Jackson JSON PropertyFilter, which selects properties of an object/list/map using a subset of the Facebook Graph API filtering syntax.</description>
-    <groupId>com.github.bohnman</groupId>
+    <groupId>com.github.jacquant</groupId>
     <artifactId>squiggly-filter-jackson</artifactId>
-    <version>1.3.19-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
-        <maven-bundle-plugin.version>3.5.1</maven-bundle-plugin.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <!-- VERSIONS -->
+        <maven-bundle-plugin.version>5.1.9</maven-bundle-plugin.version>
+        <antlr4-runtime.version>4.13.1</antlr4-runtime.version>
+        <antlr4-plugin.version>${antlr4-runtime.version}</antlr4-plugin.version>
+        <apache-commons-lang3.version>3.14.0</apache-commons-lang3.version>
+        <google-guava.version>33.2.0-jre</google-guava.version>
+        <jackson-databind.version>2.17.0</jackson-databind.version>
+        <jcip-annotations.version>1.0</jcip-annotations.version>
+        <jakarta.servlet-api.version>6.1.0-M2</jakarta.servlet-api.version>
+        <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
+        <maven-source-plugin.version>3.3.1</maven-source-plugin.version>
+        <maven-javadoc-plugin.version>3.6.3</maven-javadoc-plugin.version>
+        <maven-gpg-plugin.version>3.2.4</maven-gpg-plugin.version>
+        <nexus-staging-maven-plugin.version>1.6.7</nexus-staging-maven-plugin.version>
+        <junit-jupiter-api.version>5.11.0-M1</junit-jupiter-api.version>
     </properties>
 
     <licenses>
         <license>
             <name>BSD License</name>
-            <url>https://raw.githubusercontent.com/bohnman/squiggly-filter-jackson/master/LICENSE.md</url>
+            <url>https://raw.githubusercontent.com/jacquant/squiggly-filter-jackson/master/LICENSE.md</url>
         </license>
     </licenses>
     <developers>
         <developer>
-            <name>Ryan Bohn</name>
+            <name>Antoine Jacques</name>
         </developer>
     </developers>
     <scm>
-        <connection>scm:git:git://github.com/bohnman/squiggly-filter-jackson.git</connection>
-        <developerConnection>scm:git:ssh://github.com:bohnman/squiggly-filter-jackson.git</developerConnection>
-        <url>https://github.com/bohnman/squiggly-filter-jackson/tree/master</url>
+        <connection>scm:git:git://github.com/jacquant/squiggly-filter-jackson.git</connection>
+        <developerConnection>scm:git:ssh://github.com:jacquant/squiggly-filter-jackson.git</developerConnection>
+        <url>https://github.com/jacquant/squiggly-filter-jackson/tree/master</url>
     </scm>
 
     <distributionManagement>
