@@ -1,12 +1,12 @@
-package com.github.bohnman.squiggly.examples.dropwizard.web;
+package com.github.jacquant.squiggly.examples.dropwizard.web;
 
-import com.github.bohnman.squiggly.examples.dropwizard.model.Issue;
+import com.github.jacquant.squiggly.examples.dropwizard.model.Issue;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
 @Path("/issues")
 @Produces(MediaType.APPLICATION_JSON)
