@@ -1,12 +1,12 @@
-package com.github.bohnman.squiggly.examples.servlet.web;
+package com.github.jacquant.squiggly.examples.servlet.web;
 
-import com.github.bohnman.squiggly.examples.servlet.model.Issue;
-import com.github.bohnman.squiggly.examples.servlet.util.Jackson;
+import com.github.jacquant.squiggly.examples.servlet.model.Issue;
+import com.github.jacquant.squiggly.examples.servlet.util.Jackson;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class IssueServlet extends HttpServlet {
