@@ -1,11 +1,11 @@
-package com.github.bohnman.squiggly.examples.springboot;
+package com.github.jacquant.squiggly.examples.springboot;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.bohnman.squiggly.Squiggly;
-import com.github.bohnman.squiggly.examples.springboot.web.ListResponse;
-import com.github.bohnman.squiggly.web.RequestSquigglyContextProvider;
-import com.github.bohnman.squiggly.web.SquigglyRequestFilter;
+import com.github.jacquant.squiggly.Squiggly;
+import com.github.jacquant.squiggly.examples.springboot.web.ListResponse;
+import com.github.jacquant.squiggly.web.RequestSquigglyContextProvider;
+import com.github.jacquant.squiggly.web.SquigglyRequestFilter;
 import com.google.common.collect.Iterables;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -14,7 +14,7 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 @SpringBootApplication
 public class Application {
