@@ -1,9 +1,9 @@
-package com.github.bohnman.examples.springdatarest;
+package com.github.jacquant.examples.springdatarest;
 
 import org.hibernate.annotations.Cascade;
 import org.hibernate.annotations.CascadeType;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.List;
 
 @Entity
