@@ -1,4 +1,4 @@
-package com.github.bohnman.squiggly.examples.dropwizard;
+package com.github.jacquant.squiggly.examples.dropwizard;
 
 public class IssueConfiguration extends io.dropwizard.Configuration {
 }
